requirements.txt
asgiref==3.3.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-crispy-forms==1.12.0
django-import-export==2.5.0
django-tinymce4-lite==1.8.0
django-versatileimagefield==2.1
et-xmlfile==1.1.0
jsmin==2.2.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
psycopg2-binary==2.8.6
python-decouple==3.4
python-magic==0.4.24
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
